requirements.txt
Django
djangorestframework
django-filter
pytest
pytest-django
python-dotenv
psycopg2-binary
django-cors-headers